tests: Reduce flakiness of daemon ssl files change test
The test is changing certificates when the ovn-nbctl is running
as daemon and expects that the mismatch will be detected. In most
cases this is fine because the main loop of the ovn-nbctl will be
woken up by something external. However, when the host is busy it might
not be woken up in time.

Wake the daemon right away by calling show command. At the same time
ensure that there is only 1 mismatch after the whole process.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>