northd: Use proper field for lookup_nd
We are intentionally skipping ND NA with LLA as source.
However, this doesn't work when the ND NA has LLA source,
but the target address is global one. In that case we
would skip update of already existing entry when
always_learn_from_arp_request is set to false.

Use ND target address in the check instead of source.

Fixes: 5e0cb03605ea ("northd: Add logical flow to skip GARP with LLA")
Reported-at: https://issues.redhat.com/browse/FDP-283
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>