northd: Reduce the number of logical flows for ECMP symmetric reply
There are 3 flows for matching ECMP symmetric reply that are different
in a single match part that is the protocol (udp/tcp/sctp) for ct.new
traffic and additional 3 flows (udp/tcp/sctp) for ct.est && !ct.rpl.

Collapse those 6 flows into 1 by using or condition. One thing to
note is that in the end the number of OpenFlows will remain the same.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>