tests: Fix flaky "options:requested-chassis ...".
The test was failing due to the following race condition.
By using ofport-request, there is the risk that the ofport is initially
not assigned the expected value, and hence that the ofport changes.
This happens for instance if the geneve interface gets the ofport expected for
the vif. In that case, both the geneve and the vif see their ofports changing.
An ofport change for the geneve interface is not I+P handled. If this change
is handled within the same loop as the iface-id setting for the vif, then a
runtime data recompute happened for the vif, and the expected log was not produced.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>