northd: Don't skip transit switch LSP when creating mcast groups.
Currently when we enable IGMP on OVN-IC cluster with two or more AZs
and one vm from AZ1 send IGMP report, northd will create the following
multicast_group on each AZ:

AZ1:
 1. multicast_group that forward the mcast traffic from LS1 to the VM.
 2. multicast_group that forward the mcast traffic from LR1 to the LS1.

AZ2:
 1. multicast_group that forward the mcast traffic from TS to LR1 in
    AZ1.

This design works fine if we have one logical network only on each AZ,
but if we have two or more logical network on the same AZ that separated
from each other and only connected via transit switch and both join the
same mcast network, the traffic will be delivered between those two
networks because ovn floods it to all routers that connected to the
transit switch (see logical flow table 27).

The above design is not the right way to handle such mcast traffic
because future changes for ovn that make it match explicitly on the igmp
group address and not forward traffic to routers can break the mcast
traffic in ovn-ic.

This patch updates the above design by adding the router port that
connects to the transit switch to the multicast_group even if the peer
port have mcast_flood enabled.

Signed-off-by: Mohammad Heib <mheib@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>