controller: Allow br-int connection via other methods.
The br-int connection is hardcoded to use unix socket, which requires
for the socket to be visible for ovn-controller. This is achievable in
container by mounting the socket, but in turn the container requires
additional privileges.

Add option to ovn-controller that allows to specify remote target for
br-int. This gives the user possibility to connect to br-int in different
manner than unix socket, defaulting to the unix socket when not specified.
In addition, there is an option to specify inactivity probe for this
connection, disabled by default.

Reported-at: https://issues.redhat.com/browse/FDP-243
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>