tests: Fix netcat 7.94 issues.
The netcat 7.94 allows multiple connections over udp (-k/--keep-open) [0],
without this option the connection can be closed "unexpctedly". This
to keep the test backward compatible make new servers for every UDP
connection.

The second issue is that netcat is attempting to listen on IPv4 when
the there isn't any server address specified and fails to do so. Add
-6 flag to indicate that this is pure IPv6 connection.

[0] https://github.com/nmap/nmap/commit/4e6c8feb153c0c9ff8a68cd841669d650319ab45
Fixes: 40136a2f2c84 ("northd: Fix direct access to SNAT network.")
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>