Revert "northd: Don't skip transit switch LSP when creating mcast groups."
This reverts commit 85ca2b75369c9a73f4750d5914666a54ebb3f2e0.

The commit above breaks inter-AZ IP multicast for the case when one of
the multicast receivers is co-located in the same zone as the sender.
In those cases traffic is not correctly forwarded to other receivers
that joined the group in other AZs.

This is often the case in ovn-kubernetes deployments (with IC enabled).
The current "interconnection - IGMP/MLD multicast" unit test failed to
cover such topologies.

CC: Mohammad Heib <mheib@redhat.com>
Fixes: 85ca2b75369c ("northd: Don't skip transit switch LSP when creating mcast groups.")
Reported-at: https://issues.redhat.com/browse/FDP-656
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>