northd: Detect if SB LB was removed and re-sync that row.
When someone removes SB LB it won't be detected until the next
recompute of the "sync_to_sb_lb" node. This will cause traffic
disruptions in case of hairpin as the flows directly depend on
the SB LB entries. Add check to trigger recompute when we detect
that the row is missing in SB, but still present in NB.

Reported-at: https://issues.redhat.com/browse/FDP-682
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>