tests: Do not use check in VXLAN mode disable test.
The test tries to setup 4097 LSs via single ovn-nbctl command,
this is fine however it might lead to "Unreasonably long Xms poll
interval" on slower systems thus failing the test. Remove the check
as the test still checks if all switches were created right under
so it should be completely fine and this should avoid faliures
on slower systems.

Fixes: 7e99500e60bf ("northd: Add support for disabling vxlan mode.")
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>