tests: Compare IP and recompute related ports.
Compare related-lports as obtained by IP and after recompute,
automatically, for all unit tests.

In current OVN, a datapath is not removed from the set of local datapath
when it should (e.g. the last local port is removed).
This is a known limitation of current implementation.
Avoid test failing indirectly due to this reason (port considered as related
as on a datapath erroneoulsy seen as local).

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>