Support selection fields for ECMP routes
This patch adds the ability to specify a custom set of packet headers
for hash computation for ECMP routes similar to the support that was
added for LB in 5af304e7478adcf5ac50ed41e96a55bebebff3e8

ECMP routes by default use dp_hash as a selection_method for OVS flows.
When ecmp_selection_fields is specified, the selection_method will be
hash with the specified list of fields used for computing the hash.

For simplicity, list of fields that are used in the select action
is a union of all the fields specified in each Logical_Route_Static_Route
that is part of a given ECMP route.

Signed-off-by: Karthik Chandrashekar <karthik.c@nutanix.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>