Documentation: Add inclusive-language documentation.
A recent series of commits to OVN made some changes to the language
used in code, comments, and documentation to be more inclusive. This
constitutes a follow-up that clarifies in the documentation what the
policy is and where the list of words to avoid can be found.

When starting this task, I found a commit to OVS [1] written by Simon
Horman that added basically everything I wanted to say. With Simon's
permission, I have copied the contents of that commit, changing all
instances of "Open vSwitch" to "OVN". Simon is credited as a co-author
on this commit.

[1] https://github.com/openvswitch/ovs/commit/df5e5cf4318a

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Co-authored-by: Simon Horman <horms@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>