utilities: Moved ovn-sbctl show command from OVS.
Transferred all logic of the show command from ovs/lib/db-ctl-base.c
to ovn-sbctl.c. This is necessary in the future for the implementation
show of separate chassis. These functions cannot be imported, due to
the fact that in ovs/lib/db-ctl-base.c in these functions the global
for the module structure cmd_show_tabeles is used, which is imported
from ovn-sbctl, for filtering individual chassis I consider it
necessary to take this block separately to ovn.

Signed-off-by: Alexandra Rukomoinikova <arukomoinikova@k2.cloud>
Tested-by: Ivan Burnin <iburnin@k2.cloud>
Signed-off-by: 0-day Robot <robot@bytheb.org>