ci: Add timeout command to test execution.
The timeout command allows us to specify a timeout for each job in CI.
This is handy because there might be a test that is not able to
finish fro various reasons, last time it was forking nc process.
At the same time having timeout over the make command allows us
to preserve artifacts even if the timeout is reached so it might
be easier to debug what is happening.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>