controller: Prevent crash when db is empty.
ovn-controller crashed when connected to an empty local db, forcing the user to
initialize a local db before starting the controller daemon. With this fix, the
controller will no longer crash in this case, but instead keep running and wait
until the database is initialized to do anything else.

Reported-at: https://issues.redhat.com/browse/FDP-715
Signed-off-by: Rosemarie O'Riorden <rosemarie@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>