handle IGMP group changes in incremental processing.
To improve ovn-northd perfomance when processesing changes to IGMP
groups this patch impliments a handler to process IGMP group changes
instead of relying on a full database recompute.

I tested using a script that is avalible here:
https://gist.github.com/JacobTanenbaum/da70c75178bec2b8e1cf5487c74466e3

when running this script inside the ovn sandbox it creates sw0 and 5000
ports on sw0. It then adds one IGMP group with 1 port in it and waits
for the databases to sync

using upstream main

Time spent on processing nb_cfg 2:
	ovn-northd delay before processing:	5ms
	ovn-northd completion:			5ms

this is a typical output after adding the IGMP group waiting for the sb
to sync

while using a branch with this patch:

Time spent on processing nb_cfg 2:
	ovn-northd delay before processing:	1ms
	ovn-northd completion:			1ms

this is the time that it takes to wait for the database to sync.

Additionally the time it takes to add an IGMP group increases the more
ports are on a switch from the current head of main while on a branch
with this patch the time to add an IGMP group remains the same
regardless of the size of the database (up to the 100000 ports on the
logical switch that I have tested).

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>
Suggested-by: Dumitru Ceara <dceara@redhat.com>
Reported-at: https://issues.redhat.com/browse/FDP-756
Signed-off-by: 0-day Robot <robot@bytheb.org>