northd: Move connected routes to route engine.
previously connected routes (so routes based on the networks of the
LRPs) where handled separately from static routes.

In order to later announce routes from ovn to the fabric we need to have
a central overview over all routes we want to announce. This includes
routes based on Logical_Router_Static_Routes as well as the networks on
Logical_Router_Port.

With this change the "routes" engine node will output a complete list of
all routes in the system. This also serves as a future integration point
to integrate routes learned from outside of OVN.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>
Signed-off-by: 0-day Robot <robot@bytheb.org>