Allow LR to send RAs with only link local Ipv6.
This fixes a bug in OVN that causes ovn-controller to fail when Logical
Router Port configures send_periodic=true, but the Logical Router itself
doesn't have any globally routable IPv6 networks.

This is supported by RFC 4861 [0], in sections 6.2.2 and 4.2

This is part of a larger effort to get BGP unnumbered working within OVN.
Backport specific patch for v24.03.4.

[0]: https://datatracker.ietf.org/doc/html/rfc4861

Signed-off-by: MJ Ponsonby <mj.ponsonby@canonical.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>