ovn-controller dns-cache: Improve dns caching using cmaps.
Its possible for pinctrl handling for DNS packets to be blocked
if the ovn-controller main thread has acquired the dns cache
mutex during the recompute.  This patch improves the DNS packet
handling by using cmap instead of hmap to maintain the dns cache.
With this, there is no need to maintain a mutex.

Suggested-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>