northd: Add node for IGMP and Multicast data.
Add new I-P node that will store all the data for IGMP and
Multicast groups. This node allows to avoid full recompute of lflow
node when IGMP or Multicast SB table changes.

The node itself still does full recompute for IGMP and Multicast
changes however this is a compromise between code complexity and
the time it takes for all lflow to be created. At the same time
thi brings the benefit of having the data available when there
is recompute of the lflow node.

As design choice there is only single lflow_ref for all IGMP
lflows, that makes them not being thread safe and only main thread
can generate them during full recompute of lflow node. This shouldn't
be an issue, because the computation of igmp lflow is pretty simple.

Reported-at: https://issues.redhat.com/browse/FDP-756
Co-authored-by: Jacob Tanenbaum <jtanenba@redhat.com>
Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>
Suggested-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>