Allow LR to send RAs through localnet port.
Modifies the rule responsible for dropping the MLF_LOCAL_ONLY packets
to only drop them if the MLF_OVERRIDE_LOCAL_ONLY bit flag is not there.

This does also include the addition of MLF_OVERRIDE_LOCAL_ONLY bitflag applied
if a router announcement is being sent from either a gateway
or distributed router.

This is part of an ongoing unnumbered BGP effort.
Backport specific patch for v24.03

Signed-off-by: MJ Ponsonby <mj.ponsonby@canonical.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>