tests: Add missing reset_pcap_file() definition to EDNS test.
When backporting f22a1ba9c127 ("Skip only OVN DNS responder packets from
OUT_ACL.") we missed the fact that on branch-22.03 reset_pcap_file() had
not been factored out as a standalone helper.  That is, commit
8f841ecbe28e ("tests: Factor out reset_pcap_file() helper.") is missing
on branch 22.03.

For simplicity, adapt the EDNS test to include its own definition of the
helper (similar to other tests on the 22.03 branch).

Fixes: f22a1ba9c127 ("Skip only OVN DNS responder packets from OUT_ACL.")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>