northd: Add "persist-established" ACL option.
On allow-related ACLs, if the ACL changes and no longer matches an
established session, then traffic will no longer automatically be
allowed. Instead, traffic on the established session will be subject
to ACL matching, and therefore the traffic may be dropped.

This behavior can be altered by setting the "persist-established"
option on allow-related ACLs. When set to true, we set a specific
bit in the ct_mark when the conntrack entry is committed. If this bit
is set, then established session traffic will always be allowed, even
if the ACL is altered to no longer match the traffic.

Upcoming commits will put in place methods so that deleting the ACL or
changing the action type on the ACL will remove the conntrack entry that
allows the established traffic.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>