controller, northd: Ensure proper register usage.
The registers used in actions and lflow created by controller are
hardcoded, which causes upgrade issues. When the register are changed
the "old" version might be still propagated by northd until upgraded.

To prevent that add flag that will be populated only by northd that
is using the "new" version of registers. The controller checks
this flag and uses proper registers in lflows/actions.

Reported-by: Xavier Simonart <xsimonar@redhat.com>
Suggested-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>