northd: Fix NULL pointer deref for NAT add_route.
The commit in the fixes tag introduced support for adding LRPs
without networks, leaving them with only the IPv6 Link Local
Address (LLA).

This is a very useful feature for routing between LRs, avoiding
waste of a scarce resource such as IPv4 addresses.

It did however miss a couple of direct references to
router_port->lrp_networks.ipv4_addrs[0].

Check for presence of IPv4 addresses, and use IPv6 LLA when none
are present.

Expand "ovn -- NAT and Load Balancer flows" test case to include
a neighbour LR with LRP without IPv4 address, as that is enough
to expose the crash.

Note that more elaborate flow checks will be added in a subsequent
patch that fixes a logic bug in the routing IPv4 over IPv6 next
hop patch.

Cc: MJ Ponsonby <mj.ponsonby@canonical.com>
Fixes: 0ee90e29eb8b ("Allow creation of a LRP without ipv4.")
Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>