ci: containers: Use Ubuntu 22.04 in GitHub CI for container build stability.
We recently started having very hard (impossible?) to debug failures of
the containers CI job when building ARM images:

https://github.com/ovn-org/ovn/actions/runs/12979653009

The CI jobs start ARM Fedora or Ubuntu containers that run on top of an
Ubuntu host (using qemu user virtualization).  In GitHub actions, when
the underlying runner is using the GitHub Ubuntu 24.04 image, random
crashes seem to happen: we've seen gcc segfault seemingly without
reason; we've seen make segfault also without an obvious reason;
we've seen ninja segfault also without reason.

Due to the fact that these failures happen only in GitHub CI, it's very
hard to debug them.  I've tried reproducing this locally in a similar
setup to what I'm guessing GitHub actions use but I didn't manage to.

Given that we don't have these issues when the GitHub runner uses an
Ubuntu 22.04 image, downgrade the version of the containers job runner
to 22.04.  Ubuntu 22.04 standard support ends in Apr 2027 so that gives
us more room to investigate what the problem is on Ubuntu 24.04 or even
upgrade to the next LTS, 26.04.

NOTE: the resulting container images will still be using the same base
image as before, this change is transparent for them.  That means the
actual OVN CI tests should not notice any behavior change due to this
commit.

Suggested-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>