controller: Avoid unnecessary recomputes due to bfd_chassis.
bfd_chassis node was recomputed when sb_global columns such as nb_cfg
were updated. If, in addition, SB was read only at that moment (i.e. an
transactiuon was in fly), then the update was cancelled and resulted
in a full recompute.

This was highlighted as causing random failures in tests such as
"ACL with Port Group conjunction flow efficiency".

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>