mac-cache: Fix expiration of active FDB entry due to skipped update.
ovn-controller doesn't update timestamps until 3/4 of the expiration
threshold have passed, but then it may wait for another 3/4.  And since
3/4 + 3/4 > 1, northd would remove the entry before ovn-controller
wakes up to refresh it next time.

Fix the issue the same way it is fixed for MAC bindings - by using
cooldown period of 1/4 and the dump period of 1/2 of the threshold.

Fixes: 551527a5e68e ("controller: Update FDB timestamp")
Reported-at: https://issues.redhat.com/browse/FDP-1132
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>