ic: Fix denylist for IPv6 with same prefix length.
Fix the prefix filter function as the return condition when IPv6
prefixes have same length. If denylist prefix and prefix verified
have the same length, it must be have compared only.
Without this fix if denylist filter has
2003:db08::/64 and it exists a route to 2003:db88, this route
is blocked because the calc for IPv6 doesn't check if they have
the same prefix lenght.
AND operator between prefix and denylist prefix will be
2003:db08::/68 in this example (2003:db88::/64 & 2003:db08::64)

Signed-off-by: Lucas Vargas Dias <lucas.vdias@luizalabs.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>