northd: Fix broadcast of all traffic within a transit spine switch.
Currently, OVN can't learn FDB entries from remote ports, because
learning is happening in the ingress pipeline, while only the egress
pipeline is executed in the destination availability zone.  So, the
destination zone can't learn the source MAC address and has to
broadcast replies.  The source zone also can't learn from replies,
so has to broadcast all the subsequent packets as well.

Fix that by introducing FDB learning in the egress pipeline for
remote ports.

Note: For some reason remote ports were not tracked by ovn-controller.
This wasn't a big problem because we didn't use inport/outport matches
on the traffic from remote ports in most cases.  However, FDB lookups
need to match on the input port, and ovn-controller will skip those
flows if ports are not related, so marking them as such. (The removal
is already handled in a generic way.)

Fixes: 67100f0ca819 ("ic: Add support for spine-leaf topology for transit switches.")
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>