northd: Add LR option to commit all traffic.
Add LR option to commit all traffic that is not already committed
by either NAT or LB. This ensures that the traffic is tracked, and
we don't erroneously commit reply traffic, or reply traffic is not
marked as invalid.

To achieve the commit we need to perform lookup on every packet
that goes through LR pipeline whenever there is stateful NAT.

The SNAT lookup requires additional flag as the unSNAT is happening
in ingress pipeline and at that point we need to know if the packet
is reply or not. This is not required for DNAT, because unDNAT stage
happens in egress.

Reported-at: https://issues.redhat.com/browse/FDP-787
Signed-off-by: Ales Musil <amusil@redhat.com>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>