treewide: Added support for port mirroring in OVN overlay.
Previously, the mirroring feature in OVN only supported tunneling
traffic to local and remote ports outside the OVN cluster.

With this update, it is now possible to mirror traffic passing through
a virtual port to a dedicated OVN port.
Changes introduced in the Northbound DB schema and XML configuration:
1)new field for mirroring rules has been added to the Mirror table,
allowing the definition of traffic reflection rules.
2)new table, Mirror Rule, has been created to filter remote
overlay traffic.
3)new mirroring type, lport, has been introduced to encapsulate
mirrored traffic to another OVN port.
4)new port type, mirror, has been added to serve as a conduit to the
target port.
5)ls_in_mirror and ls_out_mirror stages in ls pipeline.

lport mirroring details:
When attaching an lport mirror to a logical switch port, a new
port binding named mp-datapath-target-port is created.
This port acts as a mirror port, with its parent being
the target port.

VLAN tagging is not required—packets are transmitted without VLAN headers.

For lport mirroring, logical flows (flows) are generated to handle
traffic processing based on filtering criteria.

Packets matching these criteria are duplicated and sent to the
target port, while the original traffic continues along its designated path.

If an lport mirror is created without specifying any rules,
default logical flows are added to mirror all incoming
and outgoing traffic to the target port.

NOTE: this feature is experimental and may be subject to
removal/change in the future.

Signed-off-by: Alexandra Rukomoinikova <arukomoinikova@k2.cloud>
Signed-off-by: Vladislav Odintsov <vlodintsov@k2.cloud>
Co-authored-by: Vladislav Odintsov <vlodintsov@k2.cloud>
Tested-by: Ivan Burnin <iburnin@k2.cloud>
Signed-off-by: 0-day Robot <robot@bytheb.org>