en-advertised-route-sync: Don't assume that switch peers are routers.
A peer of a switch can be another switch, so the port type has to be
checked.  The missing check doesn't seem to lead to crashes, but it
leads to addresses of switch-switch ports not being advertised.

Fix that by checking the port type instead of assuming.

The issue is a result of two features being developed at the same time
and the code not being re-checked before merging the routing patches.

Fixes: 93f541f342f9 ("northd: Allow announcing individual host routes.")
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>