controller: Process more SB data in one iteration.
Run the idl loop for the southbound database until there is no more data
to be received or a timeout is reached.

This improves processing of changes when controller iterations take a
long time (e.g., we see times of up to ~20s). In such a case, it is
better to collect as many changes as possible to be processed in a
single run. Previously, changes would be delayed very often because they
were stuck in the incoming queue.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>
Signed-off-by: Martin Morgenstern <martin.morgenstern@cloudandheat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>