containers: Get sparse from the official GitHub mirror.
We had a few issues lately with GitHub Actions being unable to checkout
sparse from the git.kernel.org:

  git clone git://git.kernel.org/pub/scm/devel/sparse/sparse.git
  Cloning into 'sparse'...
  fatal: unable to connect to git.kernel.org:
  git.kernel.org[0: 172.105.4.254]: errno=Connection timed out
  git.kernel.org[1: 2600:3c04::f03c:95ff:fe5e:7468]: errno=Network is unreachable

Let's try and use the official mirror of the project hosted on GitHub
itself.  This mirror is maintained by the main maintainer of the
project and mentioned in the documentation as one of the ways of
getting sparse:
  https://sparse.docs.kernel.org/en/latest/#getting-sparse

It may also be better to not create extra load on kernel.org servers,
they should not be used for CI purposes.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Acked-by: Alin Gabriel Serdean <aserdean@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>