ovn-controller: Reserve zones for upcoming ports.
This change will be useful for migration cases, where it can be
used to sync ct-entries before port is up on new chassis,
resulting in reduced network package drops.
It also fulfills the need of any other service which might need
advance ct-zone reservation in future.

Restarting this patch again which was reviewed here originally:
https://mail.openvswitch.org/pipermail/ovs-dev/2024-September/417457.html

Signed-off-by: Mansi Sharma <mansi.sharma@nutanix.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>