controller: (nit) Avoid setting again same register.
Before the patch, LOG_CT_ZONE might have been set twice, when virtual or
localports are present.

For virtual ports, for which a ct_zone is allocated when recomputing but not
in incremental processing, this also resulted in differences between flows
produced by recompute and incremental processing.

A future patch will be checking the differences between recompute and
incremental processing produced flows in all tests.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>