tests: Avoid adding two similar load balancers.
Avoid using two similar load-balancers such as:
ovn-nbctl lb-add lb4-no-aff [[fd11::a]]:8081 [[fd11::2]]:80
ovn-nbctl lb-add lb40-no-aff [[fd11::a]]:8081 [[fd11::2]]:80

Using such load balancers on the same switch would result in similar openflows
in table OFTABLE_CHK_LB_HAIRPIN, containing learn actions, and differing only
by the cookie. One of the two flows will be installed in OVS by OVN (e.g. the flow
with the cookie from lb4-no-aff).
Any flows learned by that learn action will be deleted by OVS when the
corresponding flow in table OFTABLE_CHK_LB_HAIRPIN is deleted.

In case of recompute, there is no guarantee that the same openflow is installed:
the flow with the cookie from lb40-no-aff might now be installed. This would result
in the deletion of the learned flow in OVS.

If one of the two similar load balancers is deleted, the same behaviour might happen:
the learned flow in table OFTABLE_CHK_LB_HAIRPIN_REPLY might be deleted, depending
of which load balancer was preferred.

Describing the behaviour so that, if one considers that this is a bug and that mutiple
similar load balancers can be added and result in no flow deletion during recompute,
a JIRA should be entered.

THis wa identified by comparing flows from IP and recompute and frequent failures of test
"load balancing affinity sessions - IPv6".

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>