controller, northd: Add command to enable time warp.
The time/warp can be useful in tests where we need to test time
sensitive operations, and it might not be convenient to wait
for the time to pass. The warp adjusts only monotonic clock
so it can be used only on operations that do not relay on wall
clock time.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>