tests: Use scapy contrib BFD implementation.
Due to a change of behavior in scapy 2.6.x, the current CoPP system
test no longer works.

One of the changes we need to do is to switch to using sendp()
instead of send(), which requires adding Ethernet frames.

I discovered that scapy 2.6.x includes a BFD contrib module, so
while we're already updating the scapy code, let's switch to using
the BFD class in the same breath.

To ensure we catch such regresssions in the future, unpin scapy
in the CI container image.

Reported-at: https://launchpad.net/bugs/2104222
Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>