inc-proc-eng: Streamline ENGINE_NODE macros.
In current OVN, an engine_node has two function callbacks that are
optional, is_valid() and clear_tracked_data(). There are individual
macros that allow for each of these optional functions to be defined in
the engine_node. With two optional callbacks, this means there are four
possible macros:

- vanilla engine node (no is_valid(), no clear_tracked_data())
- has clear_tracked_data(), does not have is_valid()
- has is_valid(), does not have clear_tracked_data()
- has both is_valid() and clear_tracked_data()

In OVN we have bespoke macros for the first, second, and fourth
scenarios [1].

The main issues with this are:
- This does not scale well if engine_nodes need new optional callbacks.
- The macro names are quite long.

This patch offers a new alternative. It alters the ENGINE_NODE macro so
that it can take a variable number of arguments. Some preprocessor magic
uses the number of arguments passed in to call a specific variant
(ENGINE_NODE2, ENGINE_NODE3, or ENGINE_NODE4 currently,
since these are the possible number of inputs to ENGINE_NODE). Depending
on the variant called, additional optional elements of the resulting
engine_node can be set. Convenience macros are provided for the two
optional callbacks.

Under the hood, there's a lot going on. But from the engine node
developer's perspective, it's much easier to declare engine nodes, since
there is only one macro to declare an engine node, and its arguments
help define which optional callbacks are defined.

[1] There is one engine_node (acl_id) that falls into the third
scenario but rather than use a custom macro, it just defines its
is_valid() callback after the initial engine_node definition.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>