inc-proc-eng: Remove NAME_STR argument from ENGINE_NODE variants.
With the exception of two (pflow_output and lflow_output in
ovn-controller), every engine_node in OVN uses the stringified form of
the engine_node struct name as its NAME_STR argument.

At this point, it's clear that this NAME_STR argument is redundant, so
we may as well remove it and shrink the code down by a bit.

A positive side effect of this change is that it takes the guesswork out
of searching for an engine node input by name.

The exception to this rule is the DB-related nodes. They still
differentiate between the engine node struct name (e.g. "sb_chassis")
and the engine node name string (e.g. "SB_chassis"). This isn't strictly
necessary, but keeping it this way makes the diff much much smaller :)

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>