inc-proc-eng: Make engine_set_node_state() private.
The previous two commits removed all instances where
engine_set_node_state() was called outside of inc-proc-eng.c. State
changes are all handled by the incremental engine now, so there is no
need to expose the state change function in its header.

There is a downside to this, which is that we lose visibility into where
exactly the state change was made. Debug logs previously pinpointed the
line and file where the state change was being made. This is not as
useful since state changes now only happen on a few different lines in
inc-proc-eng.c.

This change introduces a "reason" parameter for engine_set_node_state()
so that additional details can be passed in. This way, the debug message
can still contain useful information about the "why" behind the state
change, even if it can't be traced to a specific line within a node's
run() or an input's change_handler() callbacks.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>