lib: northd: Add a new ct-state-save feature flag.
This feature flag informs ovn-northd whether all hypervisors that are
part of the OVN cluster support the new ct_state_save() logical action.

A subsequent patch will use this new action to avoid generating OpenFlow
rules that cannot be offloaded to hardware on specific NICs.  This flag
is needed in order to avoid minor-version upgrade issues (if
ovn-controller is upgraded after ovn-northd).

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>