controller: Fix ct zone assignment I+P.
Previously local ports that had an external_id:iface-id set would get a
ct-zone assigned on recompute, independent if there is a matching
southbound port_binding for them or if this port_binding can actually
reside on the local chassis.
However during incremental processing this was not the case and only
bound local ports would have their ct zone added or removed (as only
they are included in tracked_dp_bindings).

To fix this we add a status enum to local_lports that defines if there
is a matching port_binding or not.
The ct_zones engine node no only considers local_lports that are backed
by some port_binding entry.

This means we can also get rid of workarounds in testcases that where
introduced in 9f052bdb because of this problem.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>
Signed-off-by: 0-day Robot <robot@bytheb.org>