northd: Introduce exclude-router-ips-from-garp option for lsp.
If this option is set on the logical switch port of type router,
then the IPs of its peer router port is excluded from the
nat_addresses so that ovn-controller doesn't send the GARPs
for these IPs.

This option will be useful for the deployments where the SNAT ips
configured on the routers is different from the router port IP
and no external entity wants to reach to the router port IPs.
In a highly scaled environment, this reduces the load on the
pinctrl thread.

CMS should not set this option if the router port IPs are also
used for SNAT.

Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>