ic: Advertise and learn routes for loadbalancer VIPs.
In an interconnect topology, when a logical router (LR) is connected to a
transit-switch (TS), the static routes on the LR as well as routes for the
connected subnets get advertised to the other sites via the IC-SB.

This commit adds two new nb_global options ic-route-adv-lb and
ic-route-learn-lb, which will be help propagate the LB VIP based routes.
Default value for both the options is false. Also these options come into
effect only if the exisiting options ic-route-adv and ic-route-learn are
true.

ic-route-adv-lb: If this is set to true, add routes in the IC-SB for the
VIPs associated with the LB instances linked to the logical router. These
routes would have the VIP as destination and IP of the LRP connected to TS
as next hop The origin would be set as "loadbalancer".

ic-route-learn-lb: If this is set to true, routes with origin loadbalancer
are synced from the IC-SB and applied to the LR as learned route.

This change makes services hotes behind LB in one site become accessible by
clients in other locations, via interconnect.

Signed-off-by: Sragdhara Datta Chaudhuri <sragdha.chaudhu@nutanix.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>